fix #47 webとmobile両方で写真を選択できるようにしました
